Move MSBuild registering code into its own method and call it from constructor.
