+++ Program/DocsPortingTool.cs
@@ -1,12 +1,6 @@
 ﻿#nullable enable
 using Libraries;
 using System;
-using System.Collections.Generic;
-using System.Reflection;
-using System.Runtime.Loader;
-using System.IO;
-using System.Linq;
-using Microsoft.Build.Locator;
 
 namespace DocsPortingTool
 {
@@ -25,11 +19,6 @@ namespace DocsPortingTool
                     }
                 case Configuration.PortingDirection.ToTripleSlash:
                     {
-                        // This ensures we can load MSBuild property before calling the ToTripleSlashPorter constructor
-                        VisualStudioInstance? msBuildInstance = MSBuildLocator.QueryVisualStudioInstances().First();
-                        Register(msBuildInstance.MSBuildPath);
-                        MSBuildLocator.RegisterInstance(msBuildInstance);
-
                         var porter = new ToTripleSlashPorter(config);
                         porter.Start();
                         break;
@@ -37,93 +26,6 @@ namespace DocsPortingTool
                 default:
                     throw new ArgumentOutOfRangeException($"Unrecognized porting direction: {config.Direction}");
             }
-        }
-
-        private static readonly Dictionary<string, Assembly> s_pathsToAssemblies = new Dictionary<string, Assembly>(StringComparer.OrdinalIgnoreCase);
-        private static readonly Dictionary<string, Assembly> s_namesToAssemblies = new Dictionary<string, Assembly>();
-
-        private static readonly object s_guard = new object();
-
-        /// <summary>
-        /// Register an assembly loader that will load assemblies with higher version than what was requested.
-        /// </summary>
-        private static void Register(string searchPath)
-        {
-            AssemblyLoadContext.Default.Resolving += (AssemblyLoadContext context, AssemblyName assemblyName) =>
-            {
-                lock (s_guard)
-                {
-                    if (s_namesToAssemblies.TryGetValue(assemblyName.FullName, out var cachedAssembly))
-                    {
-                        return cachedAssembly;
-                    }
-
-                    var assembly = TryResolveAssemblyFromPaths(context, assemblyName, searchPath, s_pathsToAssemblies);
-
-                    // Cache assembly
-                    if (assembly != null)
-                    {
-                        var name = assembly.FullName;
-                        if (name is null)
-                        {
-                            throw new Exception($"Could not get name for assembly '{assembly}'");
-                        }
-
-                        s_pathsToAssemblies[assembly.Location] = assembly;
-                        s_namesToAssemblies[name] = assembly;
-                    }
-
-                    return assembly;
-                }
-            };
-        }
-        internal static Assembly? TryResolveAssemblyFromPaths(AssemblyLoadContext context, AssemblyName assemblyName, string searchPath, Dictionary<string, Assembly>? knownAssemblyPaths = null)
-        {
-            foreach (var cultureSubfolder in string.IsNullOrEmpty(assemblyName.CultureName)
-                // If no culture is specified, attempt to load directly from
-                // the known dependency paths.
-                ? new[] { string.Empty }
-                // Search for satellite assemblies in culture subdirectories
-                // of the assembly search directories, but fall back to the
-                // bare search directory if that fails.
-                : new[] { assemblyName.CultureName, string.Empty })
-            {
-                foreach (var extension in new[] { "ni.dll", "ni.exe", "dll", "exe" })
-                {
-                    var candidatePath = Path.Combine(
-                        searchPath, cultureSubfolder, $"{assemblyName.Name}.{extension}");
-
-                    var isAssemblyLoaded = knownAssemblyPaths?.ContainsKey(candidatePath) == true;
-                    if (isAssemblyLoaded || !File.Exists(candidatePath))
-                    {
-                        continue;
-                    }
-
-                    var candidateAssemblyName = AssemblyLoadContext.GetAssemblyName(candidatePath);
-                    if (candidateAssemblyName.Version < assemblyName.Version)
-                    {
-                        continue;
-                    }
-
-                    try
-                    {
-                        var assembly = context.LoadFromAssemblyPath(candidatePath);
-                        return assembly;
-                    }
-                    catch
-                    {
-                        if (assemblyName.Name != null)
-                        {
-                            // We were unable to load the assembly from the file path. It is likely that
-                            // a different version of the assembly has already been loaded into the context.
-                            // Be forgiving and attempt to load assembly by name without specifying a version.
-                            return context.LoadFromAssemblyName(new AssemblyName(assemblyName.Name));
-                        }
-                    }
-                }
-            }
-
-            return null;
         }
     }
 }
